PartySenseWebFrontend/requirements.txt
django>=1.5
python-openid
django-compressor
oauth2
pillow
django-social-auth
django-braces
beautifulsoup4
lxml
html5lib
django-debug-toolbar
python-memcached
django-memcache-status
django-registration
django-registration-email
django-countries